+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Track.is_midi
+
 ## [0.1.0] - 2019-02-28
 
 ### Added